pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "efficientsam3"
version = "0.1.0"
description = "EfficientSAM3 research utilities and Stage-1 distillation pipeline"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Apache-2.0"}
authors = [
    {name = "Chengxi Simon Zeng"},
    {name = "Yuxuan Jiang"},
    {name = "Aaron Zhang"},
]
keywords = ["computer vision", "segmentation", "distillation", "sam3"]
dependencies = [
    "timm>=1.0.17",
    "numpy>=1.26.4",
    "tqdm",
    "ftfy==6.1.1",
    "regex",
    "iopath>=0.1.10",
    "typing_extensions",
    "huggingface_hub",
    "psutil",
]

[project.optional-dependencies]
stage1 = [
    "torch>=2.7.0",
    "torchvision>=0.18.0",
    "decord>=0.6.0",
    "mmengine>=0.10.4",
    "pycocotools>=2.0.7",
    "yacs>=0.1.8",
    "Pillow>=10.0.0",
    "opencv-python>=4.9.0.80",
    "scipy>=1.10.0",
    "scikit-image>=0.21.0",
    "scikit-learn>=1.3.0",
    "tensorboard>=2.12.0",
    "einops>=0.7.0",
    "hydra-core>=1.3.2",
    "submitit>=1.5.1",
    "fvcore>=0.1.5.post20221221",
    "fairscale>=0.4.13",
    "mmcv>=2.1.0",
    "pyyaml>=6.0.1",
    # PyPI name: segment-anything, import name: segment_anything
    # Used by some SAM1-compat utilities (e.g. EfficientViT SAM wrapper).
    "segment-anything",
]

# Optional dependency: only required for the SAM1-compat EfficientViT wrapper
# under sam3/backbones/efficientvit/efficientvit/sam.py
sam1 = [
    "segment-anything",
]
dev = [
    "pytest",
    "pytest-cov",
    "black==24.2.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/SimonZeng7108/efficientsam3"

[tool.setuptools.packages.find]
include = ["stage1*", "sam3*", "fewshot_adapter*"]

